LRAC-14784 analytics-batch-export-import-impl Extract common methods
