+++ modules/apps/analytics/analytics-batch-export-import-impl/src/main/java/com/liferay/analytics/batch/exportimport/internal/dispatch/executor/BaseAnalyticsDispatchTaskExecutor.java
@@ -1,0 +1,122 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.analytics.batch.exportimport.internal.dispatch.executor;
+
+import com.liferay.analytics.batch.exportimport.manager.AnalyticsBatchExportImportManager;
+import com.liferay.analytics.settings.configuration.AnalyticsConfigurationRegistry;
+import com.liferay.dispatch.executor.BaseDispatchTaskExecutor;
+import com.liferay.dispatch.executor.DispatchTaskExecutorOutput;
+import com.liferay.dispatch.executor.DispatchTaskStatus;
+import com.liferay.dispatch.model.DispatchLog;
+import com.liferay.dispatch.model.DispatchTrigger;
+import com.liferay.dispatch.service.DispatchLogLocalService;
+import com.liferay.petra.function.UnsafeConsumer;
+import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.exception.PortalException;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+
+import java.util.Date;
+import java.util.List;
+
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Riccardo Ferrari
+ */
+public abstract class BaseAnalyticsDispatchTaskExecutor
+	extends BaseDispatchTaskExecutor {
+
+	@Override
+	public void doExecute(
+			DispatchTrigger dispatchTrigger,
+			DispatchTaskExecutorOutput dispatchTaskExecutorOutput)
+		throws Exception {
+
+		if (!analyticsConfigurationRegistry.isActive()) {
+			return;
+		}
+
+		analyticsBatchExportImportManager.exportToAnalyticsCloud(
+			getBatchEngineExportTaskItemDelegateNames(),
+			dispatchTrigger.getCompanyId(),
+			getNotificationUnsafeConsumer(
+				dispatchTrigger.getDispatchTriggerId(),
+				dispatchTaskExecutorOutput),
+			getResourceLastModifiedDate(dispatchTrigger.getDispatchTriggerId()),
+			getResourceName(), dispatchTrigger.getUserId());
+	}
+
+	protected abstract List<String> getBatchEngineExportTaskItemDelegateNames();
+
+	protected UnsafeConsumer<String, Exception> getNotificationUnsafeConsumer(
+		long dispatchTriggerId,
+		DispatchTaskExecutorOutput dispatchTaskExecutorOutput) {
+
+		DispatchLog dispatchLog =
+			dispatchLogLocalService.fetchLatestDispatchLog(
+				dispatchTriggerId, DispatchTaskStatus.IN_PROGRESS);
+
+		return message -> updateDispatchLog(
+			dispatchLog.getDispatchLogId(), dispatchTaskExecutorOutput,
+			message);
+	}
+
+	protected Date getResourceLastModifiedDate(long dispatchTriggerId) {
+		DispatchLog latestSuccessfulDispatchLog =
+			dispatchLogLocalService.fetchLatestDispatchLog(
+				dispatchTriggerId, DispatchTaskStatus.SUCCESSFUL);
+
+		if (latestSuccessfulDispatchLog == null) {
+			return null;
+		}
+
+		return latestSuccessfulDispatchLog.getEndDate();
+	}
+
+	protected abstract String getResourceName();
+
+	protected void updateDispatchLog(
+			long dispatchLogId,
+			DispatchTaskExecutorOutput dispatchTaskExecutorOutput,
+			String message)
+		throws PortalException {
+
+		StringBundler sb = new StringBundler(5);
+
+		if (dispatchTaskExecutorOutput.getOutput() != null) {
+			sb.append(dispatchTaskExecutorOutput.getOutput());
+		}
+
+		sb.append(_dateFormat.format(new Date()));
+		sb.append(StringPool.SPACE);
+		sb.append(message);
+		sb.append(StringPool.NEW_LINE);
+
+		dispatchTaskExecutorOutput.setOutput(sb.toString());
+
+		dispatchLogLocalService.updateDispatchLog(
+			dispatchLogId, new Date(), dispatchTaskExecutorOutput.getError(),
+			dispatchTaskExecutorOutput.getOutput(),
+			DispatchTaskStatus.IN_PROGRESS);
+	}
+
+	@Reference
+	protected AnalyticsBatchExportImportManager
+		analyticsBatchExportImportManager;
+
+	@Reference
+	protected AnalyticsConfigurationRegistry analyticsConfigurationRegistry;
+
+	@Reference
+	protected DispatchLogLocalService dispatchLogLocalService;
+
+	private static final DateFormat _dateFormat = new SimpleDateFormat(
+		"yyyy-MM-dd'T'HH:mm:ss.SSSZ");
+
+}
